+++ src/controllers/ForbiddenWordsController.js
@@ -1,6 +1,6 @@
 var logger = require('logops');
 var ForbiddenWordDao = require('../daos/ForbiddenWordDao');
-var AdminDao = require('../daos/AdminUserDao');
+var { checkIsAdmin } = require('../helpers/RequestHelper');
 var { sendSuccessResponse, sendErrorResponse, sendEmptySuccessResponse } = require('../helpers/ResponseHelper');
 
 class ForbiddenWordsController{
@@ -24,7 +24,7 @@ class ForbiddenWordsController{
         };
         
         try{
-            await AdminDao.findByToken(req.get('adminToken'));
+            await checkIsAdmin(req);
 
             var forbiddenWord = await ForbiddenWordDao.create(data);
             logger.info(`Forbidden word '${data.word}' added to organization ${data.organizationId}`);
@@ -39,7 +39,7 @@ class ForbiddenWordsController{
         var wordId = req.params.id;
 
         try{
-            await AdminDao.findByToken(req.get('adminToken'));
+            await checkIsAdmin(req);
             
             await ForbiddenWordDao.delete(wordId);
             logger.info(`Forbidden word '${wordId}' deleted`);
